feat(runtime-core): move the runtime-core setup logic into shuvi
