+++ packages/runtime-core/package.json
@@ -15,13 +15,9 @@
     "template"
   ],
   "scripts": {
-    "dev": "run-p watch:*",
-    "watch:app": "tsc -p tsconfig.build.app.json -w",
-    "watch:node": "tsc -p tsconfig.build.node.json -w",
-    "prebuild": "rimraf lib shuvi-app",
-    "build": "run-p build:*",
-    "build:app": "tsc -p tsconfig.build.app.json",
-    "build:node": "tsc -p tsconfig.build.node.json"
+    "dev": "tsc -p tsconfig.build.json -w",
+    "prebuild": "rimraf lib",
+    "build": "tsc -p tsconfig.build.json"
   },
   "engines": {
     "node": ">= 12.0.0"
