+++ packages/runtime-core/src/application/create-application-server-spa.ts
