+++ packages/runtime-core/src/application/history/create-hash.ts
