--- packages/runtime-core/tsconfig.build.node.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "compilerOptions": {
-    "module": "commonjs",
-    "outDir": "lib"
-  },
-  "include": ["src/lib"]
-}
