+++ packages/shuvi/src/api/api.ts
@@ -21,7 +21,6 @@ import { joinPath } from '@shuvi/utils/lib/string';
 import { deepmerge } from '@shuvi/utils/lib/deepmerge';
 import invariant from '@shuvi/utils/lib/invariant';
 import { Hookable } from '@shuvi/hooks';
-import coreRuntime from '@shuvi/runtime-core';
 
 import { setRuntimeConfig } from '../lib/runtimeConfig';
 import {
@@ -29,11 +28,8 @@ import {
   normalizeRoutes,
   renameFilepathToComponent
 } from '../lib/pageRoutes';
-import {
-  PUBLIC_PATH,
-  ROUTE_NOT_FOUND_NAME
-} from '../constants';
-import initRuntime from '../initRuntime';
+import { PUBLIC_PATH, ROUTE_NOT_FOUND_NAME } from '../constants';
+import initRuntime from '../lib/initRuntime';
 import { defaultConfig, IConfig, loadConfig } from '../config';
 import { IResources, IBuiltResource, IPlugin, IPreset } from './types';
 import { Server } from '../server';
@@ -400,7 +396,6 @@ class Api extends Hookable implements IApi {
 
     // Runtime installation need to be executed before initializing presets and plugins
     // to make sure shuvi entry file at the top.
-    coreRuntime.install(this.getPluginApi());
     this.runtime.install(this.getPluginApi());
     runPlugins();
   }
